Refactor webpack-blocks/core
